requirements.txt
gradio==6.1.0
groq==0.37.1
python-dotenv==1.2.1
